Bulk partition update in RtpsRelay is inefficient
Problem
-------

A bulk update of the partitions serviced by an RtpsRelay instance has
two problems.  First, it is repetitive and wastes bandwidth since most
of the partitions will already be in the set.  Second, it is subject
to fragmentation which create additional overhead.

Solution
--------

Partition the set of partitions to create additional instances of the
topic.  The goal is that each instance is small enough to fit in a
single UDP message to avoid fragmentation but large enough to fill up
a message.