java: installation of jar files
This patch install the jar files into INSTALL_LIBDIR directory instead
the `lib` directory.
This allows to be uniform with the rest of the installation of OpenDDS
libraries.
And it removes an error when the installation uses the DESTDIR variable
and with the `lib` directory doesn't exist. This is particularly true
for the packaging of OpenDDS. The rpm packaging renames the
opendds_dcps.jar file to $(DESTIDR)$(INSTALL_PREFIX)/lib, and it is
impossible to copy the other files inside the directory.