ICE password changes increase load on RtpsRelay
Problem
-------

A participant that is using the RtpsRelay and ICE will send ICE
password changes to the application participant of the relay.  If the
participant changes to a different relay, then the participant will
changes its ICE password, send the samples, and then heartbeat.  The
new relay will not ack these heartbeats.  Relay application
participants typically have long lease durations so the unwanted
heartbeats could go on for a significant amount of time.

Solution
--------

1. Provide a PID that allow the application participant to identify
itself.

2. Provide logic that purges application participants when changing
the relay address.