Improvements to Sphinx Documentation
I've had these squirreled away away for awhile, but because of #2750 I
figured I should clean it up and put it out so it can be used the next
time we need to write/convert documentation. This will conflict with
 #2750 in a minor way, so it can be merged first.

- Added a system for linking to GitHub in various ways. Examples shown
  below this list.
- Ignore sphinx related paths in lint.pl.
- Added a workflow to check for Sphinx warnings and broken links in the
  documentation.
- Removed development guidelines from Sphinx for now because they set
  off warnings and broken links checks. Besides it was sorta meant to be
  filler content.

```rst
:ghissue:`213`
```

turns into the equivalent of:

```rst
  `Issue #213 on GitHub <https://github.com/objectcomputing/OpenDDS/issues/213>`_
```

```rst
:ghpr:`1`
```

turns into the equivalent of:

```rst
  `Pull Request #1 on GitHub <https://github.com/objectcomputing/OpenDDS/pull/1>`_
```

```rst
:ghfile:`README.md`
```

turns into the equivalent of:

```rst
``README.md``
```

that is a link to that file on GitHub. It will try to point to the most
specific version of the file:
  - If the OpenDDS is a release it will calculate the release tag and use that.
  - Else if the OpenDDS is in a git repository it will use the commit hash.
  - Else it will use `master`.