OpenDDS lacks intentional unit test coverage
Problem
-------

Intentional unit test coverages means that 1. for every unit, which is
a group of source files with the same prefix, there is a corresponding
set of unit tests and 2. the coverage report for the files belonging
to that unit are generated by executing the corresponding unit test.
The goal is that each line of code is covered by a unit that was
intending to execute the line of code.

Solution
--------

Adjust the names of the unit tests and a bash script to generate the
report.  The script can be converted to perl if we need support
non-bash platforms.