Network issues lead to rediscovery
Problem
-------

Participants that connect to the RtpsRelay over the public internet
will often have problems.  For example, a cable modem that is on the
fritz may be up for a minute and then down for 14 minutes.  This adds
load to the RtpsRelay as it has to discover the participant each time
if the participant's lease is small.

Solution
--------

Provide an option that allows a participant to extend the lease of
discovered participants.