Empty partitions degrade RtpsRelay performance
Problem
-------

The RtpsRelay is most efficient when an application is divided into a
number of small partitions.  Participants that (inadvertently) don't
specify any partitions are effectively placed in the same partition.
This partition can become large which degrades the performance of the
RtpsRelay.

Solution
--------

Add a control that allows the empty partition "" to be filtered.