Fix Wireshark Dissector with GLib 2.68
This is the GLib included with Ubuntu 21.04.

Continues work on problem discovered as part of
https://github.com/objectcomputing/OpenDDS/pull/2916.

This GLib PR https://gitlab.gnome.org/GNOME/glib/-/merge_requests/1715
added code with C++ linkage to GLib headers which we were wrapping with
`extern "C"`. Our's and Wireshark's usage of `exetern "C"` apparently
didn't line up with what GLib developers thought we should be doing.

Tested with Wireshark 1.2.18, 1.6.16, and 3.4.4. Might be broken in
other versions...