RtpsRelay does not log discovery time
Problem
-------

The time required for the relay to discover a participant is an
important datum for measuring the performance of the RtpsRelay.
Currently, this datum is not logged.

Solution
--------

Measure the time from first generic SPDP announcement to the time the
participant was discovered.

(cherry picked from commit 3cbc56ae51d6ed0279df6d35b7a2e76dd198c48f)