Only using doc-group3.
There is a rather big letter soup in and under DDS.
Having two ACE/TAO implemenations is a kind of freedom.
But having for the DOC group the vesion 6.x and 7.x is, IMHO, a bit too much.

What I understood from Johnny is that --doc-group3 is, or could be,
the most used option. [1]

This patch is update the README that wasn't mentioning the ACE 7.0 / TAO 3.0.
And removes the two switches in the configure file for the doc-group.

Feel free to apply or not for a next release.

[1] https://github.com/objectcomputing/OpenDDS/commit/fd6813eaff7a29aadea45d6278628266b0348b6b#commitcomment-55427969

Signed-off-by: Jan Vermaete <jan.vermaete@gmail.com>