Fix DevGuide Messenger CMakeLists Header Cleanup
Fixes issue seen in
https://github.com/objectcomputing/OpenDDS/issues/3017.

The CMakeLists file is supposed to clean up generated headers in the
source because they can conflict with the CMake build. It couldn't do
that for two reasons:
  - There was a typo in the code: `_headers` instead of `headers`. This
    made the code non-functional.
  - If the typo is fixed it would remove `DataReaderListenerImpl.h`
    because the `file(GLOB` paths are absolute and the code the excludes
    the header isn't.

Fixed the typo and changed it to remove the result of `file(GLOB
"DataReaderListenerImpl.h")` instead of just
`"DataReaderListenerImpl.h"` from the list headers to remove.