Create privacy-policy.html
We need to add a privacy policy page to the site, which the updated support terms & conditions will link to. The meta data and html may need to be adjusted to make this work, but it should be pretty close to ready to publish. Once the page is live, we will need to add a link to it from the footer. Let me know if you have questions.