RtpsRelay:  Application participant guid used in public interface
Problem
-------

Most of the RtpsRelay topics are keyed by the application participant
guid.  To monitor a specific instance, one must first use the Relay
Instance topic to find the relay with the correct user data, extract
the application participant guid, and then use it to filter results.

Solution
--------

RtpsRelay instances will be identified by user-provided string.  A
status topic has been added for monitoring the health of an RtpsRelay
instance.