Fix `OPENDDS_TARGET_SOURCES`/Qt MOC CMake Conflict
Fixes https://github.com/objectcomputing/OpenDDS/issues/2963

The combination of using `add_dependencies` on a single target and the
Qt generated code target was causing CMake to error because of cycles in
the target dependency graph. This changes it so we don't do that if the
targets are the same.

Also:
- Add CMakeLists.txt to the Qt-based ishapes demo to test this change.
- Fixed warning about IDL headers being evaluated by Qt.
- Fixed `find_package(Threads)` apparently needing C to be enabled in
  newer CMakes if a C/C++ project hasn't been declared yet.