ReactorTask catches exceptions
Problem
-------

The main loop in ReactorTask catches exceptions.  When an exception is
thrown, the main loop will log the exception (good) and terminate the
thread without killing the entire process (bad).  Killing the entire
process provides instant feedback that something is wrong and needs to
be corrected.

Solution
--------

Rethrow the exception.