`remove_domain_participant` may cause deadlock
Problem
-------

Suppose a listener that acquires a lock is installed for the builtin
topics.  Suppose that a thread acquires the lock before calling
`remove_domain_participant`.  This will deadlock since the listener
will be invoked synchronously and try to acquire the lock.

Solution
--------

Invoke the listener for BITs asynchronously.

For reference, compare the approach taken for ConnectionRecords.