RemoveAssociationSweeper is unnecessary
Problem
-------

The RemoveAssociationSweeper class was designed to allow the
"reader (to) process more information that may have queued up from the
writer."  This queueing of data could occur in the transport or the
DataReader.  If the author was referring the DataReader, then the
RemoveAssociationSweeper is superfluous because the DCPS half of the
DataReader is already providing the reader process an opportunity to
process samples.  If the author was referring to the transport, then
the transport is at fault for not promptly deliverying the samples to
the DataReader.

Solution
--------

Remove the RemoveAssociationSweeper.