SporadicTask optimizations are ineffective, buggy, and complicated
Problem
-------

There are two optimizations for SporadicTask:

1. The first optimization attempts to only submit a schedule command
   if not scheduled or the scheduled dispatch time is nearer than the
   desired dispatch time.

   This logic is controlled by variables that are only set by the
   reactor thread.  Thus, a non-reactor thread attempting to schedule
   could still create a large number of ScheduleCommands before the
   reactor thread gets around to processing the first one.

   The optmization that examines the next dispatch time changes the
   semantics of SporadicTask in that it is implied that the user can
   schedule a dispatch nearer in time without canceling.  With the
   introduction of code that modulates a SporadicTask, this created a
   bug where calls to cancel the timer were dropped and/or not
   written.  By not canceling the timer, the `scheduled_` flag was
   never reset meaning that scheduling for an earlier dispatch never
   occurred.

2. The second optmization attempts to reduce the number of calls into
   the reactor itself.  This took the form of hooks around the
   lifecycle of a ReactorInterceptor::Command.  For SporadicTask, the
   goal of the implementation was to skip commands that were
   effectively no-ops.

Solution
--------

The new design of SporadicTask is based on the idea that there is a
desired state provided by the user (scheduled or not scheduled) and
there is an actual state which is the result of interacting with the
reactor.  When scheduling or canceling, the same
ReactorInterceptor::Command is submitted to the reactor.  When
executed, the command attempts to make the actual state the desired
state if they differ.  To avoid dynamic allocation, the command is
allocated once and only put on the ReactorInterceptors queue if it is
not on the queue already.

Notes
-----

The pattern for SporadicTask should be adopted by PeriodicTask,
MultiTask, and other uses of the ReactorInterceptor.