`set_listener` is not atomic
Problem
-------

There is a race condition where the listener for an entity can be
invoked after the user sets the listener to nil.  The spec may allow
this behavior, however, it leads to behaviors that are
counterintuitive.  Specifically, consider the problem of using a
listener with a read condition.  A typical shutdown sequence might be
1) set the listener to nil, and 2) remove the read condition.  Due to
the race condition, the listener may be invoked and use a read
condition that is no longer available.

Solution
--------

Make `set_listener` atomic with respect to invoking the listener.