QueryCondition test has race condition on listener
Problem
-------

The listener in the QueryCondition test uses a read condition.  The
shutdown sequence in the test is:

1. `set_listener`
2. `delete_readcondition`

Currently, OpenDDS may retain a reference to and invoke the listener
after the call to `set_listener`.  In the problematic sequence, the
listener is being invoked after the call to `delete_readcondition`
which causes the error.

Solution
--------

Add logic to disable the listener before it is unset.