Asserting liveliness prolongs lifetime of all instances
Problem
-------

Assume there are two writers on a topic using
`MANUAL_BY_TOPIC_LIVELINESS_QOS` and `assert_liveliness` and assume that
each instance belongs to a single writer.  If one writer goes away,
the instance state for the instances that it writes does not go to
`NOT_ALIVE_NO_WRITERS_INSTANCE_STATE` because the liveliness message
of the other writer prolong its lifetime.

Solution
--------

When applying a liveliness update to an instance, check that the
writer actually writes that instance.