ICE::Agent singleton accessed during shutdown
Problem
-------

Programs that exit normally will call the destructor for the
ICE::Agent singleton.  Later in the destruction sequence, the
singleton may be accessed again and re-created.  This leads to
problems since the newly created ICE::Agent cannot perform the desired
actions.

Solution
--------

Save a handle to the ICE::Agent instead of always going through the singleton.