"Listener" interfaces cause complexity
Problem
-------

Various components in OpenDDS need to share changes in state with
other components.  Let one component be the Publisher and the other
component be the Subscriber.  The Publisher defines a listener
interface.  The Subscriber implements the listener interface and
subscribes to changes in the Publisher.  When the Publisher wants to
share a change in state, it *synchronously* invokes the Subscriber
through the listener interface.

The problem with this design is the opacity of virtual method calls
and concurrency.  The thread that invokes the Publisher will be
holding some locks.  These locks will still be held when the
Subscriber is invoked.  Most likely, the Subscriber will acquire locks
when invoked.  Thus, the Subscriber needs to be designed in a way that
avoids deadlock.  This is very brittle since any change on the code
path that invokes the Publisher must be analyzed with respect to every
Subscriber.  This is intractable from a maintenance perspective.

One way of dealing with the problem is to defer processing in the
Subscriber.  In this design, the Subscriber posts a job to a queue and
processes the change with a clear view of what locks are held.

Solution
--------

Making the typically solution more generic leads to an "Internal DDS."
An InternalDataWriter writes a sample.  The sample is placed in an
InternalDataReader's cache and an InternalDataReaderListener is
scheduled for execution.  InternalTopics allow InternalDataReaders to
find InternalDataWriters.