OpenDDS lacks directions for creating a minimum working example
Problem
-------

The preferred method of reporting a bug in OpenDDS is to create a
minimum working example.  However, there are no directions for
creating and submitting a minimum working example.

Solution
--------

Update the README with directions for the minimum working example and
provide a small "Hello World!" example as a starting point.