DataReaderImpl_T allocator is not thread safe
Problem
-------

Allocations from `data_allocator_` in DataReaderImpl_T are
synchronized on the `sample_lock_`.  However, not all deallocations
are not synchronized.  For example, taking samples will place the
samples in a sequence and destroying the sequence may cause a
deallocation.

Solution
--------

Use a mutex in the relevant part of the allocator.