`and_wait` causes deadlock
Problem
-------

Thread A has a lock L and is deleting a writer.  The transport thread
B that services the writer has just received a sample and is invoking
a listener that acquires L.  The logic for deleting a writer invokes
various `and_wait` methods to stop periodic and sporadic tasks.  This
deadlocks because the reactor thread (B) that must execute the
commands is blocked on the lock L.

Solution
--------

* Remove `and_wait` functions.

* Use weak pointers to avoid dangling references.

Note
----

In general, waiting on a command should be removed as a feature.  This
can be done in many cases with appropriate conversions to Sporadic or
Periodic Task.  In all other cases, the Command submitted to the
ReactorInterceptor should be refactored to use weak pointers or a new
helper class that uses weak pointers should be devised.