TransportRegistry has load data race
Problem
-------

A thread calling `create_inst` only waits if a load is pending.  For a
static build, there is no guarantee that the thread loading the
static libraries will have started the load.  Thus, `create_inst`
fails since it doesn't wait and then can't find the transport type.

Solution
--------

If desired, wait for the transport to be loaded by checking
`type_map_`.