Sample lock is held too long
Problem
-------

Setting the instance state may lead to invoking the listener.  To
allow the user to read samples, the sample lock is unlocked with a
reverse lock.  The reverse lock only works if the sample lock is
locked once.  (It is a recursive lock which is a problem in and of
itself.)  Some code paths that set the instance state lock the sample
lock multiple times so the reverse lock doesn't release the lock.

The problem is that the transport thread can be storing a sample and
then invoking the listener.  At this point the sample lock is not
locked.  Another thread, i.e., discovery or the service participant,
can then update the instance state due to liveliness.  This thread
will acquire and retain the sample lock when invoking the listener.
If the transport thread does a read or take, which it probably will,
it will deadlock.

Solution
--------

Remove the extra locks.