`writer_became_dead` may lead to deadlock
Problem
-------

The call into `instances_liveliness_update` will increment the count
of the recursive sample lock.  This can eventually result in a listener being
invoked.  The sample lock must not be held when the listener is
invoked because it creates the opportunity for user code to deadlock.

Solution
--------

Apply proper scoping of the sample lock.