Multicast association has numerous problems
Problem
-------

Adding a remote starts a timer to send SYN packets.  Upon receipt of
the SYN_ACK, the timer is cancelled.  The timer cannot be re-enabled.
This makes the following sequence problematic:

1. Participant A creates a writer.

2. Association with a remote reader completes.

3. Participant B creates another writer that should associate with the
   same reader.

The backoff for the syn timer also backs off without reset.  Thus, a
new reader could be penalized by a older reader.

Solution
--------

* Remove DataLinkWatchdog and replace it with SporadicTask.

* Reset the syn timer each time a new reader is added.