Support INSTALL_LIB in CMake Module
Follow up to https://github.com/objectcomputing/OpenDDS/pull/2740

Support `INSTALL_LIB` being used with `make install` to change the name
of the `lib` directory. To make it possible for `find_package` to find
the OpenDDS CMake module in this case, the installation destination for
the CMake module was changed from `lib` to `share`.