GitHub Actions: default.features needed in ACE_TAO_u18_stat_qt_ws_sec
OpenDDS's configure when given an already-configured ACE+TAO will not
edit the default.features.  We need default.features to have the expected
features in order to generate makefiles for specific tests later on.

This means that the generated ACE_TAO artifact is biased for how we're
going to use it in OpenDDS, but it's only used once so this should be fine.