Json parsing requires "end of document"
Problem
-------

JsonValueReader can be used to load data.  The format of such an input
file should be a sequence of JSON Objects.  Currently, this doesn't
parse correctly as rapidjson expects there to be exactly one object in
the stream.

Solution
--------

Update the parse flags to stop after one object and not check for end
of document.