Fix 2 Anon. sequence<string> in the same struct
Defining two anonymous string sequence members in the same struct causes
a redefinition compile error in the type support. This happens because
the anonymous type code relies on the pointer to the AST_Decl of the
base type to tell if the type has already been used. The problem with
that is apparently there can be multiple objects for string types.

This replaces the pointer with the classification of the container type
and the classification and name of the base type.