key information not passed to `begin_struct_member`
Problem
-------

ValueWriter implementations may treat members that are keys
differently.  As a concrete example, consider a ValueWriter that
prepares a SQL INSERT statement for a received sample.  Most likely,
this will be an UPSERT which often requires a list of keys and a list
of non-keys.

Solution
--------

Change `begin_struct_member` to take an XTypes::MemberDescriptor which
contains the member name and a bool for key status.  This PR only sets
these two members.