Fix One of the Install Test Configure Commands
A warning added in CMake 3.23 warns when there's an unused argument to
CMake. This is the case here where this argument is supposed to be a
list being passed to a child CMake command. It's getting interpreted as
a separate argument instead. The solution is, apparently, to replace it
with this generator expression.