Duplicate samples are received in reflective mode
Problem
-------

In reflective mode, the RTPS UDP transport attempts to ignore nacks
that are pipelined due to latency.  This is done by sampling the
heartbeat count to determine a minimum for the count member of an
acknack.  In the current code, this limit was updated every time a new
enough acknack was received.

In this scheme, there is the opportunity to send a heartbeat
immediately after an acknack is received but before the nack
response.  This means the acknack that reflects this heartbeat would
be processed.  Most likely, it would be a duplicate of the acknack
that caused the nack response.

Solution
--------

Adjust the limit in the nack response.  This will cause the writer to
ignore acknacks until the reader has had an opportunity to receive any
requested data.