Instance-level security is not implemented
Problem
-------

The DDS security specification specifies functions for checking that
local or remote writers can register or dispose an instance.
Currently, these methods are defined in the appropriate interfaces but
they are not called.

To complicate matters, the functions are defined using DynamicData.
Thus, a prerequisite is to build up support for DynamicData where a
custom plugin for instance-level security is viable.

Solution
--------

* Build up support for DynamicData.

* Add calls to check when registering/disposing instances.

* Added a function for exposing the subject name was added for writing
  custom plugins.

Notes
-----

This conversion from the existing DynamicType and DynamicData
implementation was difficult due to differences in memory management.
Specifically, the DynamicData interface uses ValueTypes which have
some rough edges.  Reviewing the rules for ValueTypes is recommended.
For example, if the parameter is reference counted (interface type or
valuetype) then the callee must decrement the reference if
reassigning.