Associations not removed when participant expires
Problem
-------

Suppose that a local participant is using the RTPS transport and that
a remote participant's lease duration is less than the passive connect
duration.  Suppose that discovery proceeds to the point where a reader
on the local is sending preassociation acknacks to a remote writer
when the writer terminates.

The local reader will continue to send preassociation acknacks because
the call to remove the remote writer never comes.

Solution
--------

Add a call to `stop_accepting_or_connecting` when a pending
association is removed.