StaticDiscovery shutdown looks like startup
Problem
-------

Due to lazy loading in the transport framework, StaticDiscovery
"simulates" heartbeats and acknacks to drive discovery.  That is,
StaticDiscovery declares local writer/remote reader and local
reader/remote writer pairs to the RtpsUdpDataLink.  The datalink sends
heartbeats and acknacks for local writers and readers.  When activity
from the remote is detected (which may be simulated), the association
is added and the local writer/reader is created.

When a writer or reader is deleted, the data link resumes this
behavior, i.e., it sends heartbeats for writers that don't exist and
acknacks for readers that don't exist.  These messages can causes the
association to be added again.  For tests that rely on the
publication/subscription match status, this presents a problem since
they will need to wait an additional cycle (10 times the heartbeat
period) to shutdown.

Solution
--------

Do not send acknacks for readers that don't exist.