Status flag race leads to missing listener invocation
Problem
-------

Invocation of listeners for builtin topics is asynchronous.  Before,
invoking the listener, it checks for the DATA_AVAILABLE_STATUS flag
and then resets it after the listener is invoked.  This creates a race
condition where a new sample could be inserted before the listener
completes.  At that point, the status is incorrect and the listener
will not be called.

Solution
--------

Clear the status flag before invoking the listener.