RtpsUdpSendStrategy sporadically reports failure to encode submessage
Problem
-------

The RtpsUdpSendStrategy sporadically reports an error when encoding a
submessage.  The submessage kind is a heartbeat so it must be a
writer.  Review of the code shows that it is probably the volatile
writer.

Solution
--------

Improve the logging to confirm it is the volatile writer and narrow
down the root cause.  The two possible causes are:

1. Sending to `GUID_UNKNOWN` which is a bug in the RtpsUdpDataLink.

2. No crypto handle for the remote datareader.  This could be caused
   by a race condition where the heartbeat (presumably caused by
   association) is sent before the handle is in place.

Also, improve the atomicity in the Crypto plugin.