SPDP refuses to shutdown
Problem
-------

Tests that create and destroy participants very quickly sometimes
block.  The root cause is that the registration of handlers is
asychronous.  The bad sequence is that the unregistration of handlers
happens before the registration.  Registering the handler increases
the reference count of the SpdpTransport which blocks shutdown.

Solution
--------

Do not register the handlers if shutdown has started.