OpenDDS applications hang when dynamic library loading fails
Problem
-------

A user that fails to set the path to the OpenDDS shared libraries will
experience a deadlock when starting an application.  As part of #3342,
a condition variable was added that allows one thread to wait while
another thread loads the required transport type.  In the case of the
missing path, the wait will wait forever leading to the deadlock.

Solution
--------

* Preemptively load all of the shared libraries known by OpenDDS.
  This is currently done in the initialization of the Service
  Participant.

* Remove the condition variable.

Note
----

For dynamic libraries, it is theoretically okay for a
transport-like-thing, e.g., InfoRepoDiscovery, to depend on another
transport-like-thing, e.g., TCP transport, because the load order of
the libraries can be adjusted to accommodate the dependency.  However,
this may not work for static libraries because the load order is
indeterminate.  Thus, it may be necessary to refactor
InfoRepoDiscovery so its static initializer doesn't depend on TCP
being loaded.