SEDP deadlocks on shutdown
Problem
-------

The thread call `remove_inst` on the TransportRegistry acquires the
lock and then calls `shutdown` on the instance.  This thread
eventually wants the reactor token to shutdown the reactor.

The reactor thread may be processing an event (has the reactor token)
that causes the clean up of SEDP.  This causes SEDP to call
`remove_config` on the TransportRegistry.  `remove_config` wants the
lock.

Solution
--------

Call `shutdown` on the transport inst without the lock.