RtpsRelay doesn't have embedded webserver for meta discovery
Problem
-------

Clients of an RtpsRelay must be configured with the SPDP, SEDP, and
DATA ports of an RtpsRelay instance.  This can be done statically or
dynamically.  In the dynamic scenario, it is convenient to use
TCP/HTTP load balancing to select an instance.  The instance can
return its ports in response to an HTTP request from the client.  For
load balancing, it is useful to include a health check endpoint that
is tied to the RtpsRelay admission control.

The HTTP server to provide this functionality can be implemented
external to the RtpsRelay.  However, deployments will be simplified by
including an embedded webserver in the RtpsRelay can serve the
configuration and answer the healthcheck.

Solution
--------

Embed a webserver in the RtpsRelay that can respond to configuration
and healthcheck requests.